+++ crates/store/re_protos/proto/rerun/v1alpha1/frontend.proto
@@ -2,12 +2,13 @@ syntax = "proto3";
 
 package rerun.frontend.v1alpha1;
 
+import "google/protobuf/duration.proto";
 import "google/protobuf/timestamp.proto";
 import "rerun/v1alpha1/catalog.proto";
 import "rerun/v1alpha1/common.proto";
 import "rerun/v1alpha1/manifest_registry.proto";
-import "rerun/v1alpha1/tasks.proto";
 
+// TODO: what becomes of this comment?
 // TODO(cmc): I've decided to re-use the underlying response types for now, in the name of expediency.
 // This is bad in the long term (breaking changes to the internal APIs should not result in breaking
 // changes to the public API and, in general, the two being coupled doesn't make much sense).
@@ -16,6 +17,7 @@ import "rerun/v1alpha1/tasks.proto";
 // at the moment.
 // We can revisit this as we start getting out of `v1alpha1`.
 
+// TODO: this is a horrible name for our product's API
 // Redap's public API.
 service FrontendService {
   rpc Version(VersionRequest) returns (VersionResponse) {}
@@ -114,14 +116,15 @@ service FrontendService {
   rpc ScanTable(ScanTableRequest) returns (stream ScanTableResponse) {}
 
   // --- Tasks service ---
+
   // Query the status of submitted tasks
-  rpc QueryTasks(rerun.redap_tasks.v1alpha1.QueryTasksRequest) returns (rerun.redap_tasks.v1alpha1.QueryTasksResponse) {}
+  rpc QueryTasks(QueryTasksRequest) returns (QueryTasksResponse) {}
 
   // Fetch the output of a completed task
-  rpc FetchTaskOutput(rerun.redap_tasks.v1alpha1.FetchTaskOutputRequest) returns (rerun.redap_tasks.v1alpha1.FetchTaskOutputResponse) {}
+  rpc FetchTaskOutput(FetchTaskOutputRequest) returns (FetchTaskOutputResponse) {}
 
   // Query the status of submitted tasks as soon as they are no longer pending
-  rpc QueryTasksOnCompletion(rerun.redap_tasks.v1alpha1.QueryTasksOnCompletionRequest) returns (stream rerun.redap_tasks.v1alpha1.QueryTasksOnCompletionResponse) {}
+  rpc QueryTasksOnCompletion(QueryTasksOnCompletionRequest) returns (stream QueryTasksOnCompletionResponse) {}
 
   // -- Utilities --
 
@@ -354,4 +357,67 @@ message DoMaintenanceRequest {
   // the value of `cleanup_before` timestamp even if it's more recent than 1 hour.
   // ⚠️ Do not ever use this unless you know exactly what you're doing. Improper use will lead to data loss.
   bool unsafe_allow_recent_cleanup = 5;
+}
+
+// --- Tasks ---
+
+// A task is a unit of work that can be submitted to the system
+message Task {
+  // Unique identifier for the task
+  rerun.common.v1alpha1.TaskId id = 1;
+  // Type of the task
+  string task_type = 2;
+  // Task-type dependant data necessary to de-serialize the task
+  bytes task_data = 3;
+}
+
+// `SubmitTasksRequest` is the request message for submitting tasks
+message SubmitTasksRequest {
+  repeated Task tasks = 1;
+}
+
+// `SubmitTaskResponse` contains, for each submitted task
+// its submission outcome, encoded as a `RecordBatch`
+message SubmitTasksResponse {
+  rerun.common.v1alpha1.DataframePart data = 1;
+}
+
+// `QueryTasksRequest` is the request message for querying tasks status
+message QueryTasksRequest {
+  // Empty queries for all tasks if the server allows it.
+  repeated rerun.common.v1alpha1.TaskId ids = 1;
+}
+
+// `QueryTasksResponse` is the response message for querying tasks status
+// encoded as a record batch
+message QueryTasksResponse {
+  rerun.common.v1alpha1.DataframePart data = 1;
+}
+
+// `QueryTasksOnCompletionRequest` is the request message for querying tasks status.
+// This is close-to-a-copy of `QueryTasksRequest`, with the addition of a timeout.
+message QueryTasksOnCompletionRequest {
+  // Empty queries for all tasks if the server allows it.
+  repeated rerun.common.v1alpha1.TaskId ids = 1;
+  // Time limit for the server to wait for task completion.
+  // The actual maximum time may be arbitrarily capped by the server.
+  google.protobuf.Duration timeout = 2;
+}
+
+// `QueryTaskOnCompletionResponse` is the response message for querying tasks status
+// encoded as a record batch. This is a copy of `QueryTasksResponse`.
+message QueryTasksOnCompletionResponse {
+  rerun.common.v1alpha1.DataframePart data = 1;
+}
+
+// `FetchTaskOutputRequest` is the request message for fetching task output
+message FetchTaskOutputRequest {
+  // Unique identifier for the task
+  rerun.common.v1alpha1.TaskId id = 1;
+}
+
+/// `FetchTaskOutputResponse` is the response message for fetching task output
+message FetchTaskOutputResponse {
+  // The output of the task, encoded as a record batch
+  rerun.common.v1alpha1.DataframePart data = 1;
 }
